Broke out the getGoalCandidates and shuffleArray functions from main for better organization. #28
