+++ client/utils/get-goal-candidates.js
@@ -1,0 +1,15 @@
+const { isHidden, isHiddenCorner } = require('./is-hidden.js')
+
+function getGoalCandidates(board) {
+  let candidates = []
+  for (let i = 0; i < board.length; i++) {
+    for (let j = 0; j < board[i].length; j++) {
+      if (isHidden(board[i][j]) && !isHiddenCorner(board[i][j])) {
+        candidates.push(board[i][j])
+      }
+    }
+  }
+  return candidates
+}
+
+module.exports = getGoalCandidates
